run_flux/flux_evaluate/requirements.txt
numpy>=1.21.0
Pillow>=8.0.0
scikit-image>=0.18.0
tqdm>=4.60.0

